renaming setup-preinstall-rbac commas
