+++ image/cli/mascli/mas
@@ -827,14 +827,14 @@ case $1 in
     mas-cli setup-rbac "$@"
     ;;
 
-  setup-preinstall-rbac)
+  pre-install)
     echo "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!" >> $LOGFILE
-    echo "!! setup-preinstall-rbac                                                     !!" >> $LOGFILE
+    echo "!! pre-install                                                               !!" >> $LOGFILE
     echo "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!" >> $LOGFILE
-    # Take the first parameter off (it will be "setup-preinstall-rbac")
+    # Take the first parameter off (it will be "pre-install")
     shift
-    # Run the new Python-based setup-preinstall-rbac
-    mas-cli setup-preinstall-rbac "$@"
+    # Run the new Python-based pre-install command
+    mas-cli pre-install "$@"
     ;;
 
   gitops)
