+++ python/src/mas/cli/install/app.py
@@ -71,7 +71,7 @@ from mas.devops.tekton import (
     testCLI,
     launchInstallPipeline
 )
-from mas.devops.preinstall_rbac import applyPreInstallMASRBAC, permissionCheckForRBAC
+from mas.devops.pre_install import applyPreInstallMASRBAC, permissionCheckForRBAC
 
 logger = logging.getLogger(__name__)
 
@@ -139,19 +139,19 @@ class InstallApp(BaseApp, InstallSettingsMixin, InstallSummarizerMixin, ConfigGe
                 f"You selected the '{self.getParam('mas_permission_mode')}' permission mode.",
                 "The pre-install RBAC required for this permission mode has not been applied by your current cluster login.",
                 "This step must be completed by an OpenShift cluster administrator before MAS installation can continue.",
-                "Ask your OpenShift administrator to run 'mas setup-preinstall-rbac' for this MAS instance, MAS version, permission mode, and selected apps.",
+                "Ask your OpenShift administrator to run 'mas pre-install' for this MAS instance, MAS version, permission mode, and selected apps.",
                 "If that has already been done, you can continue the installation without applying it again."
             ])
 
-            if not self.yesOrNo("Has your OpenShift administrator already run 'mas setup-preinstall-rbac' for this installation"):
-                self.fatalError("Installation aborted. Ask your OpenShift administrator to run 'mas setup-preinstall-rbac' for this installation and then run mas install again.")
+            if not self.yesOrNo("Has your OpenShift administrator already run 'mas pre-install' for this installation"):
+                self.fatalError("Installation aborted. Ask your OpenShift administrator to run 'mas pre-install' for this installation and then run mas install again.")
         else:
             self.fatalError(
                 "\n".join([
                     f"You selected the '{self.getParam('mas_permission_mode')}' permission mode.",
                     "The pre-install RBAC required for this permission mode has not been applied by your current cluster login.",
                     "This step must be completed by an OpenShift cluster administrator before MAS installation can continue.",
-                    "Ask your OpenShift administrator to run 'mas setup-preinstall-rbac' for this installation and then rerun 'mas install' with --skip-preinstall-rbac."
+                    "Ask your OpenShift administrator to run 'mas pre-install' for this installation and then rerun 'mas install' with --skip-preinstall-rbac."
                 ])
             )
